Refactor the reconciliation logic out of pkg/gitops
